+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # @codesync/react
 
+## 2.2.5
+
+### Patch Changes
+
+- add crypto and logger
+
+## 2.2.4
+
+### Patch Changes
+
+- add logger
+
+## 2.2.3
+
+### Patch Changes
+
+- 43dc37a: compile in 2 formats
+
 ## 2.2.2
 
 ### Patch Changes